# Conflicts:
#	MeetupPlatformApi/Domain/Meetup.cs
#	MeetupPlatformApi/Features/Meetups/GetMeetup/MeetupInfoDto.cs
#	MeetupPlatformApi/Features/Meetups/GetMeetups/GetMeetupsFeature.cs
#	MeetupPlatformApi/Features/Meetups/GetMeetups/MeetupInfoDto.cs
#	MeetupPlatformApi/Persistence/EntitiesConfiguration/MeetupEntityConfiguration.cs
#	MeetupPlatformApi/Persistence/Migrations/ApplicationContextModelSnapshot.cs

+++ MeetupPlatformApi/Features/Meetups/DeleteMeetup/DeleteMeetupFeature.cs
@@ -1,7 +1,9 @@
 ﻿namespace MeetupPlatformApi.Features.Meetups.DeleteMeetup;
 
+using MeetupPlatformApi.Authentication.Helpers;
 using MeetupPlatformApi.Persistence.Context;
 using MeetupPlatformApi.Seedwork.WebApi;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -18,7 +20,9 @@ public class DeleteMeetupFeature : FeatureBase
     /// </summary>
     /// <response code="204">If deleting was successful.</response>
     /// <response code="404">If needed meetup is null.</response>
+    /// <response code="403">If organizer is trying delete meetup that he didn't create.</response>
     [HttpDelete("/api/meetups/{id:guid}")]
+    [Authorize(Roles = Roles.Organizer)]
     [ProducesResponseType(StatusCodes.Status204NoContent)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     public async Task<IActionResult> DeleteMeetup([FromRoute] Guid id)
@@ -27,6 +31,11 @@ public class DeleteMeetupFeature : FeatureBase
         if (meetup is null)
         {
             return NotFound();
+        }
+
+        if (meetup.OrganizerId != CurrentUser.UserId)
+        {
+            return Forbid();
         }
 
         context.Meetups.Remove(meetup);
